give libxml a bytestream
if the python libxml fromstring method sees an <xml>-encoding attribute, it tries to parse the string based on that encoding.
this is how to make it work.

hint: there might also be the better solution of
stripping that attribute under the assumption that the encoding in the
vim buffer is correct. This works for me(tm); will revisit later.